requirements_minimal.txt
# 엣지 배포용 최소 패키지
# Azure AutoML 모델 로드를 위한 핵심 패키지만

# Core Azure ML
azureml-core==1.61.0.post1
azureml-automl-core==1.61.0.post1
azureml-training-tabular==1.61.0

# Data & ML
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.5.1
scipy==1.10.1
joblib==1.2.0

# Forecasting essentials
lightgbm==4.6.0
statsmodels==0.13.5
xgboost==1.5.2

# Azure dependencies (최소)
azure-core==1.36.0
azure-identity==1.17.0
msrest==0.7.1

# Others
python-dateutil==2.9.0.post0
pytz==2024.2
